// develop-frontend/src/api/generated/models/IslandResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IslandResponse = {
    id: string;
    name: string;
    location: string;
    pricePerNight: number;
    description: string;
    images: Array<string>;
    amenities: Array<string>;
    isAvailable: boolean;
    lat?: number;
    lng?: number;
};